Check ESLint and rename components from storybook.
