+++ src/components/HelloWorld.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 defineProps<{
   msg: string
-}>()
+}>();
 </script>
 
 <template>
