+++ src/stories/AppPage.vue
@@ -54,20 +54,20 @@
 </template>
 
 <script lang="ts" setup>
-import './page.css';
-import MyHeader from './Header.vue';
+import "./page.css";
+import MyHeader from "./AppHeader.vue";
 
-import { ref } from 'vue';
+import { ref } from "vue";
 
 const user = ref<{ name: string } | null>(null);
 
 const onLogin = () => {
-  user.value = { name: 'Jane Doe' };
+  user.value = { name: "Jane Doe" };
 };
 const onLogout = () => {
   user.value = null;
 };
 const onCreateAccount = () => {
-  user.value = { name: 'Jane Doe' };
+  user.value = { name: "Jane Doe" };
 };
 </script>
