+++ src/stories/MyButton.vue
@@ -3,8 +3,8 @@
 </template>
 
 <script lang="ts" setup>
-import './button.css';
-import { computed } from 'vue';
+import "./button.css";
+import { computed } from "vue";
 
 const props = withDefaults(defineProps<{
   /**
@@ -18,7 +18,7 @@ const props = withDefaults(defineProps<{
   /**
    * size of the button
    */
-  size?: 'small' | 'medium' | 'large',
+  size?: "small" | "medium" | "large",
   /**
    * background color of the button
    */
@@ -27,14 +27,14 @@ const props = withDefaults(defineProps<{
 }>(), { primary: false });
 
 const emit = defineEmits<{
-  (e: 'click', id: number): void;
+  (e: "click", id: number): void;
 }>();
 
 const classes = computed(() => ({
-  'storybook-button': true,
-  'storybook-button--primary': props.primary,
-  'storybook-button--secondary': !props.primary,
-  [`storybook-button--${props.size || 'medium'}`]: true,
+  "storybook-button": true,
+  "storybook-button--primary": props.primary,
+  "storybook-button--secondary": !props.primary,
+  [`storybook-button--${props.size || "medium"}`]: true,
 }));
 
 const style = computed(() => ({
@@ -42,7 +42,7 @@ const style = computed(() => ({
 }));
 
 const onClick = () => {
-  emit("click", 1)
+  emit("click", 1);
 };
 
 </script>
